requirements.txt
cssselect
lxml
PyYaml
requests
treq
Twisted
